mocks for app auth dao moved into separate file (mock_daos.go -> mock_dao_application_authentication.go)
